Handle missing logentry tags in CommonClient.log_default
It's possible for XML logentry tags to be missing child tags, such as
msg. I've seen this in one of our organization's repositories, but it's
also confirmed by the schema definition. In fact, all the child tags of
logentry are optional; see line 29 of subversion/svn/schema/log.rnc in
the Subversion source code:
http://svn.apache.org/viewvc/subversion/trunk/subversion/svn/schema/log.rnc?view=markup

When this happens, log_default throws a KeyError when trying to
access the tag from a dictionary it's built, and therefore stops
yielding messages.

This change uses dict.get() instead of square brackets, so if a tag is
missing, the value in the yielded namedtuple will be None.