Raise SvnCommandError when the SVN CLI command returns an error code.
This allows the caller to make a distinction between this specific error,
or any other cause of a ValueError. This is especially important since
run_command() is a generator, so the exception is going to be raised when
iterating over the result (rather than at call time).

SvnCommandError is a subclass of ValueError in order to stay backward
compatible with existing exception handlers.