Add changelist to status entries
Fix files in changelists don't appear in status(). Add changelist
attribute to status() result.

Since `svn status` includes files in changelists, it makes sense for
status() to include them as well. Add a field to _STATUS_ENTRY for
changelist. Files without a changelist have changelist=None and
otherwise changelist is the name of the changelist.

Rework the status test to use assertIsNotNone and assertEqual to give
better error messages. Invalid filenames produce KeyError and are not
None, so checking for None wasn't really preventing errors.