Return empty dictionary when file has no properties
Fix #89: properties method failed when no props defined

When a file has no properties, the xml doesn't contain 'target' and
target_elem is None. To conform to the documentation for properties,
return an empty dict (instead of None).

Test for both no properties and mime type property (writing a binary
file in the simplest way I can find that works in py2 and py3).